+++ NEWS.md
@@ -1,4 +1,4 @@
-# dplyr (development version)
+# dplyr 1.0.3
 
 * `summarise()` no longer informs when the result is ungrouped (#5633).